renamed scss files to match conventions
